refactor ContentFooter
